[refactor] 공통 네비바 현재 path에 따라 아이템 활성화
