+++ src/shared/lib/index.ts
@@ -1,4 +1,3 @@
-export { navItems } from './navigation'
 export { tw, cn } from './tw'
 
 // 데이터 포멧 function
