+++ src/shared/ui/navbar/Navbar.stories.tsx
@@ -1,9 +1,10 @@
 'use client'
-import { useState, useEffect } from 'react'
+
 import type { Meta, StoryObj } from '@storybook/nextjs-vite'
 import NavItems from './NavItems'
-import { MainNav } from '@/shared/lib/navigation'
+
 import { ChartNetwork, LayoutGrid, TextSearch } from 'lucide-react'
+import { MainNav } from '@/shared/ui/routes'
 
 const mockNav: MainNav = {
   href: '/space',
@@ -19,49 +20,19 @@ const mockNav: MainNav = {
 const meta: Meta<typeof NavItems> = {
   title: 'Components/Navigation/NavItems',
   component: NavItems,
-  argTypes: {
-    menuState: { control: 'object' }
-  }
+  argTypes: {}
 }
 
 export default meta
 type Story = StoryObj<typeof NavItems>
 
 // Wrapper: 내부 상태 + Controls 동기화
-const Wrapper = ({
-  item,
-  menuState: menuStateArg
-}: {
-  item: MainNav
-  menuState: { main: string | null; sub: string | null }
-}) => {
-  const [menuState, setMenuState] = useState(menuStateArg)
-
-  useEffect(() => {
-    setMenuState(menuStateArg)
-  }, [menuStateArg])
-
-  const toggleMainMenu = (label: string) => {
-    setMenuState(prev => ({
-      main: prev.main === label ? null : label,
-      sub: null
-    }))
-  }
-
-  const toggleSubMenu = (label: string) => {
-    setMenuState(prev => ({
-      ...prev,
-      sub: prev.sub === label ? null : label
-    }))
-  }
-
+const Wrapper = ({ item, pathName }: { item: MainNav; pathName: string }) => {
   return (
     <div className="w-64 border border-gray-200 p-4">
       <NavItems
+        pathName={pathName}
         item={item}
-        menuState={menuState}
-        toggleMainMenu={toggleMainMenu}
-        toggleSubMenu={toggleSubMenu}
       />
     </div>
   )
@@ -72,7 +43,7 @@ export const Closed: Story = {
   render: () => (
     <Wrapper
       item={mockNav}
-      menuState={{ main: null, sub: null }}
+      pathName="/archive"
     />
   ),
   name: '메뉴 닫힘'
@@ -82,7 +53,7 @@ export const MainOpen: Story = {
   render: () => (
     <Wrapper
       item={mockNav}
-      menuState={{ main: '스페이스', sub: null }}
+      pathName="/space"
     />
   ),
   name: '메인 메뉴 열림'
@@ -92,7 +63,7 @@ export const SubOpen: Story = {
   render: () => (
     <Wrapper
       item={mockNav}
-      menuState={{ main: '스페이스', sub: '차트 보기' }}
+      pathName="/space" /* or 정확한 child href */
     />
   ),
   name: '서브 메뉴 열림'
